Initial implementation of a compartmentalised network stack.
The firewall compartment provide ingress and egress filtering, with
holes opened for each permitted connection.

The public API is not intended to be Berkeley sockets compatible.
Berkeley sockets were designed to abstract over many different
protocols and date back to a time when name resolution was an optional
add-on to networking.

The current API is limited to outbound TCP connections.