Add SNTP implementation.
This wraps FreeRTOS' CoreSNTP library.  Currently it queries
pool.ntp.org and doesn't verify signatures (both of these should be
changed for production implementations).

This also exposes a POSIX `gettimeofday()` API that uses the cycle
counter to calculate the current time based on the cycle time when the
time was sync'd with SNTP.  This will gradually drift and so SNTP can be
used to resynchronise at any point.