Add an API to receive into a preallocated buffer.
This avoids callers such as the TLS stack, which have preallocated space
for holding the received data, needing to do an extra copy.  This is
likely to be common for things written without expecting a zero-copy
API.

At the same time, factor out some code duplication in the three places
this API is needed.  This helps reduce code size.