Add MQTT example.
This initial example connects with an MQTT test broker, subscribes and
publishes on a dummy topic with QoS 1, unsubscribes from the topic, and
disconnects.  It also checks that ACKs (SUBACK - ACK for subscribes,
UNSUBACK - ACK for unsubscribes, PUBACK - ACK for publishes), and
notifications for publishes are received as expected.  Finally, it
checks that the implementation does not leak bytes from the heap between
the connect and the disconnect.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@scisemi.com>