WIP demo hacking
Some real fixes, some hacks, some exploratory things.

The demo works until we lose the connection to the server (which seems
to happen the first time it tries to send a keep-alive packet - I tried
lowering the threshold for timeouts but it doesn't seem to make a
difference).

The server seem to send some big (687-byte, larger than MSS) packets.
Not sure why.  That was causing them to be dropped because we couldn't
allocate the receive buffer when I turned on a lot of debugging things,
but it seems to work without them.