Fix double free/unclaim in `network_socket_close`.
We call `heap_free` on the socket early in the function, bailing out if
it failed. Then we call it a second time right at the end before
destroying the token. This was introduced as part of
e4ea2b2fc9e49e8f51d4b1e4985d4c68b6c7648d.

We can safely remove this line as the heap_free was done earlier.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@scisemi.com>