Detect failures to close socket in FreeRTOS.
Unlike the API reference suggests, `FreeRTOS_closesocket` *can* actually
fail if the event queue is full or the socket argument is invalid.  In
the current state of the code, this makes us silently leak the socket
allocation.

This commit fixes the potential leak by checking the result of
`FreeRTOS_closesocket` and re-ordering operations in
`network_socket_create_and_bind` and `network_socket_close`.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@scisemi.com>