Add MQTT API to generate client IDs at random.
Client IDs often need to be randomly generated, but need to follow a
number of rules (length, alphabet). Add an API to do this properly and
use it in the example.

While we are at it, validate client IDs passed to `mqtt_connect`.  The
coreMQTT library does not validate them, and sending an invalid client
ID to the broker is annoying as it leads the broker to terminate our
connection, which in turn results in an infinite retry loop if the
caller passed an infinite timeout. Since this is only there for
convenience / not security-relevant, only enable the validation in debug
mode.

Also, better document the lifetime of the client ID string.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@scisemi.com>