Detect connection death during connecting.
If the connection dies while we are connecting, the current code runs
into a tight loop that does nothing, only bounded by the timeout (which,
if set to unlimited, will simply block the program forever).

Detect death of the link and exit, so that the caller can retry a call
to `mqtt_connect`.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@scisemi.com>