Fix two issues regarding MQTT topics/filters.
The first issue was in the example: `testTopic` was changed from `char
[]` to `char *` over the development of the example, but
`sizeof(testTopic)` was not updated accordingly, leading to an incorrect
length being passed to the MQTT library.

Fix the `sizeof` issue by going for an std::string_view.

The second related issue was in the MQTT library itself, which did not
correctly specify whether or not the topic/filter parameters should be
NULL-terminated. The MQTT specification mandates that filters and topics
MUST NOT contain the NULL character. Thus, passing a NULL-terminated
string would result in the connection being terminated.

Fix this issue by documenting requirements on topics and filters in the
API, and doing some validation to ensure that misuses result in -EINVAL.
As the latter is only there for convenience / not security-relevant, we
only enable in debug mode.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@scisemi.com>