Add a `newSession` parameter to `mqtt_connect`.
MQTT clients may want to resume an existing session with the broker by
setting the `CleanSession` MQTT parameter. In this case (and only in
this case), the client can re-use an existing client ID.

Previously `mqtt_connect` only supported fresh sessions. Address this by
exposing a `newSession` parameter in the signature of `mqtt_connect`.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@scisemi.com>