Fix `network_socket_receive_internal` when second `FreeRTOS_recv` fails.
If the second call to `FreeRTOS_recv` fails,
`network_socket_receive_internal` bad things may happen.

First, we may return FreeRTOS failure codes which are not documented by
our public API. These failure codes are not supposed to be exposed, we
should catch them and convert them to our own errno-style error codes.

Probably even worse, this may lead to leaking the buffer, because the
number of bytes read would be negative or zero, and thus the caller may
assume that the buffer is invalid.

While at it, fix the documentation of
`network_socket_receive_preallocated`, which should never return 0.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@scisemi.com>