Fix free of offset-ed network buffer.
Network buffers are allocated in `pxGetNetworkBufferWithDescriptor` with
`heap_allocate(_, MALLOC_CAPABILITY, _)` and offset-ed with
`ipBUFFER_PADDING`. Later on, buffers are freed in
`vReleaseNetworkBufferAndDescriptor` with `free`.

Before freeing the buffer, we should remove the `ipBUFFER_PADDING`
offset. Although passing an offset-ed buffer to `free` may work in
practice, it is a bit wobbly.

For consistency with the allocation site above, we should also use
`heap_free` with `MALLOC_CAPABILITY` instead of `free`. Although both
are equivalent, this makes the code look a little bit more correct.

Lastly, it would be nice to check the result of `heap_free` and output a
debug message in case of failure. Any failure here will nearly always
result in OOM, which can be tricky to debug.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@scisemi.com>