Fix two errors in handling `receive_records` errors return codes.
`tls_connection_receive_internal` converts the `-ENOTCONN` returned from
`receive_records` into 0. This is annoying because the caller will not
interpret a return value of 0 as `connection closed`, potentially
blocking the system into an infinite loop.

In `tls_connection_create`, we should also consider the return value 0
of `receive_records` as an error value. Note that `receive_records`
should never return 0 in the first place, but we should better be
defensive.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@scisemi.com>