Optionally use a separate quota for the TCP/IP buffer manager.
The buffer manager is responsible for allocating network buffers, which
differs from the other types of allocations the TCP/IP stack performs.
It may thus make sense to give it its own quota.

For now, use a macro for to enable/disable this behavior. At some point
we may want to expose this through a build system configuration option.

While at it, move the free failure check of
`vReleaseNetworkBufferAndDescriptor` to a debug assert to optimize code
size, as this is never supposed to happen unless we have a bug in our
code or in FreeRTOS.

Also add a TODO - we likely want to pre-allocate or pool buffer
descriptors, as there is only a finite and known number of them.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@scisemi.com>