Work around a clang analyser bug.
The version that we're using incorrectly models the lifetimes of things
with C++ structured binding.  This small refactor generates identical
code but makes the lifetimes something that the analyser can correctly
reason about.

This commit also fixes a bug in the clang-tidy script where I'd
accidentally made it exclude everything except locks.hh, rather than
exclude a small bit of locks.hh.

Now that clang-tidy is actually running, fix the things that it warns
about.