Fix firewall policy for DHCP.
We currently filter out any DHCP reply that happens after the initial
DHCP handshake. So, when the TCP/IP stack tries to renew its DHCP lease,
it fails and falls back to its default address, which breaks everything.

We should permit any DHCP reply to be on the safe side.

This is quite permissive, unfortunately we cannot do more without
shipping a DHCP state machine in the firewall, which is undesirable.

We *could* add a check to forbid any DHCP server address change, however
that would break some legitimate situations where the DHCP server
address changed.

This bug was initially found and fixed in an experimental branch by
David.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@scisemi.com>