Fix allocator capability definition in SNTP.
The allocator capability definition was removed in
b9d3fbc9823862607216385a01811b7995b4916e and broke SNTP.

The removal likely happened because clang-tidy reports an error when we
define this capability in SNTP: "redefinition of
'__default_malloc_capability', previous definition in stdlib.h".

This error is a false positive because clang-tidy doesn't know that we
build with `CHERIOT_CUSTOM_DEFAULT_MALLOC_CAPABILITY`.

Instead of re-introducing the allocator capability definition, only
define `CHERIOT_CUSTOM_DEFAULT_MALLOC_CAPABILITY` for C files. That way,
clang tidy is happy for the C++ code, and the code is correct. We don't
run clang tidy on the C code anyways.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@scisemi.com>