Fix custom malloc quota issue in the buffer manager.
Unfortunately, an incomplete debug change was pushed as part of
f5561cd1d47d037650f47598501ef1ead5eaf134, leading to the capability
passed to `heap_free` being `MALLOC_CAPABILITY` and not the custom
malloc quota of the buffer manager. Address this.

While at it, rename `BM_MALLOC_QUOTA` into `BM_MALLOC_CAPABILITY`, as it
is too similar to `BUFFER_MANAGER_QUOTA`, the macro which defines the
size of the quota.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@scisemi.com>