Fix failure to remove firewall entries.
We do not call the firewall API consistently: when we add entries, we
pass ports in host endianness (see `network_socket_connect_tcp`),
however when we remove them we pass ports in network endianness. As a
result, upon removal, the firewall cannot find the relevant entry in the
table and fails to remove the rule.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@scisemi.com>