Restart the TCP/IP stack on crash.
The TCP/IP stack currently has a lot of state and cannot recover
gracefully in case of failure. This commit addresses this limitation.

We introduce a custom error handler for the TCP/IP stack. At a
high-level, this error handler terminates all user threads currently
present in the TCP/IP compartment and resets the IP thread to its entry
point `ip_thread_entry`. It does so by setting all the locks of the
compartment for destruction and notifying all futex waiters. Then it
frees all heap allocations and resets globals. After all the state has
been cleaned up, it restarts the network stack.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@scisemi.com>