Rework `network_socket_close` error paths.
49f42c1e1f7fcbacace0bcd52a58f98013168616 forgot to add unlocks on error
paths in `network_socket_close`. Address this by using a `LockGuard` and
the new `release` API.

At the same time, clarify that we have two types of errors here: those
where the call can be retried (e.g., called with the wrong malloc
capability, or an invalid timeout), and those which cannot be retried
(the socket close failed due to an unspecified error and we went ahead
to free it). Both are currently under -EAGAIN. We should differentiate
the two to give a chance to the caller to call again with the right
arguments. Address this by moving unrecoverable errors to
-ENOTRECOVERABLE and update the documentation accordingly.

Finally, a socket close return value of 0 is unrecoverable and most
likely due to an internal error, so in that case we should go ahead and
free everything instead of returning -EAGAIN.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@scisemi.com>