Better handle TCP/IP stack crashes in the NETAPI compartment.
When the TCP/IP stack crashes, API calls to the compartment return
`-ECOMPARTMENTFAIL`.

Currently `-ECOMPARTMENTFAIL` failures are not considered by the NetAPI
compartment. This leads to erratic behavior when the network stack
crashes, such as adding nonsensical entries to the firewall.

Unfortunately these checks are really tricky to add in the case of
`socket_network_kind` and `network_host_resolve` because they return
structs.

This commit reworks the API so that these functions instead take a
pointer to a struct, and return an integer error code.

Building on this, add all the necessary checks.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@scisemi.com>