Make SNTP connection failure (more) reliable.
There are two reliability issues in the SNTP code.

Right now if we fail in the last step of `ntp_time_update` (when we do
`Smtp_ReceiveTimeResponse`) because of a network error, we run into an
infinite loop of retries. Even worse, the loop does not print anything
(the system looks frozen).

Also, although `sntp_update` is run in a loop in the examples, we do not
reset the timeout before calling `sntp_update`, which leads subsequent
calls to always fail.

Address this by not calling again `Smtp_ReceiveTimeResponse` if we reach
the timeout and reset-ing the timeout before calling again `sntp_update`
in examples.

Signed-off-by: Hugo Lefeuvre <hlef@cs.ubc.ca>