Clarify and fix API endianness specification.
We have a number of APIs in the firewall and the netAPI which do not
specify in which endianness arguments should be provided. Further, we
have many cases where we use `ntohs` and `htons` the incorrect way
around.

Ensure that endianness of APIs is documented everywhere and conversions
are done correctly.

Also, consistently use `htons` and `ntohs`, not `FreeRTOS_htons` and
`FreeRTOS_ntohs`.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@ubc.ca>