Update NetAPI documentation to return untagged values on error.
The NetAPI currently uses null as common error value for pointer types.

This is problematic as it does not cover "compartment-wide" types of
failure, such as `-ECOMPARTMENTFAIL` or `-ENOTENOUGHSTACK`.

Address this by documenting error values for pointer types as untagged
values instead of null.

This does not require changes in the NetAPI implementation, as null is
itself untagged.

Examples do not currently check error values so do not need updating
either. (Note: maybe they should to give an example of proper use of
the API...)

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@ubc.ca>