TLS documentation: clarify return values and blocking behavior.
The TLS currently uses null as error value for pointer types, which does
not work well when the compartment fails with `-ECOMPARTMENTFAIL` or
`-ENOTENOUGHSTACK`. Document error values for pointer types as untagged
values instead of null.

This does not require changes in the TLS implementation, as null is
itself untagged.

Also clarify blocking behavior for each API.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@ubc.ca>