Switch to a stack-overflow-resilient error handler.
The TCP/IP stack cannot currently recover from a crash due to a stack
overflow in the TCP/IP compartment.  This is due to a limitation of the
implementation of the switcher, which cannot trigger the error handler
on stack overflow. Further, the current implementation of the error
handler runs at the point in the stack where the compartment crashed.
This implies that the error handler runs on a potentially very small
stack, making it possible that it will itself crash due to a stack
overflow.

This commit addresses these limitations using David's recently-merged
stackless error handler mechanism: https://github.com/CHERIoT-Platform/cheriot-rtos/pull/301

The error handler is now declared at thread entry points (API entry
points and driver entry point for user threads, and IP thread start for
the IP thread). We can thus eliminate the big
`compartment_error_handler`.

The new error handler runs at the top of the stack, which eliminates all
issues mentioned earlier (provided the stack is large-enough, which
should be checked through `STACK_CHECK` - a separate problem).

We keep a single error handler for user threads and the IP thread
(`reset_network_stack_state`), differentiated through a `isIpThread`
argument. Though we could separate this into two distinct error handlers
for user threads and for the IP thread, it would result in a lot of
duplicated error-prone code.

Note that we do not "reinstall context" or "force unwind" anymore in the
error handler. This is all handled under the hood by the new error
handler, and we simply exit the error handler once done, which happens
to be the place where we would leave the compartment in the force unwind
case, or the place where we would reinstall the context for the IP
thread.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@ubc.ca>