Fix bug in the error handler upon user thread crash.
If a user thread crashes and the error handler of the user thread
triggers a crash in the IP thread, the current code will block forever
because the IP thread does not release the `ipThreadLockState`.

This problem stems from the fact that we only tested user thread crashes
in a setting where the IP thread cleanly returned from `prvIPTask`
(without a crash, i.e., without triggering the IP thread error handler).

Address this by releasing the `ipThreadLockState` in both cases. Add
comments to explain.

Note: this is a pre-existing bug to the stack-overflow-resilient error
handler, so separate commit.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@ubc.ca>