Isolate the DNS resolver into its own compartment.
In the current implementation, the TCP/IP compartment is responsible for
DNS, which allows it to spoof translations to bypass connection
capabilities and create arbitrary firewall entries.

To prevent this, we must isolate the DNS resolver into its own
compartment that plugs directly at the firewall-level.

Since pulling the DNS resolver out of FreeRTOS+TCP is tricky and
existing resolver libraries are too heavyweight, we write our own one.

Similarly to the FreeRTOS+TCP resolver, this implementation is a stub
resolver: it only supports IPv4 recursive servers for A, AAAA, and CNAME
queries. This resolver also assumes that the recursive resolver recurses
into CNAME records.

The address of the DNS server is provided by the DHCP server. This may
be overly restrictive - we should evolve it in future PRs or eventually
replace it with something more featureful as the need arises.

Since the resolver plugs directly with the firewall, it must send
packets at the level of Ethernet frames - which makes it in turn
necessary to know its own IP address, the IP address of the DNS server,
its own MAC address, as well as the MAC address of the DNS server (or
that of the gateway if the server is outside of the local network). The
IP address of the DNS server and the MAC address of the device are
obtained from the firewall. The IP address of the device and the MAC
address of the server/gateway are obtained from DHCP and ARP, whose
corresponding packets are also forwarded to the DNS compartment.

Some commands which I found useful to test this code:

`arping -I ${interface used by the device} -c 1 -s ${IP of the gateway} -A ${IP of the device}`

This generates an ARP packet that sets/changes the MAC address of the
gateway. It is necessary to test the ARP-handling code if the DHCP
server has the same IP as the DNS server (in which case the MAC address
of the server will not be obtained through ARP).

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@ubc.ca>