firewall: send a read-only/non-capturable buffer to the TCP/IP stack.
The firewall currently sends a writable and capturable frame buffer
capability to the TCP/IP stack. This is bad because the TCP/IP stack can
keep the capability and alter the buffer at a later point when we re-use
it. Not sure what the exact impact is, but it sounds like the TCP/IP may
be able to use this to add endpoints to the firewall table.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@ubc.ca>