Make waiting behavior on socket locks use steps.
If a thread somehow corrupts the socket list, we will lose the ability
to retrieve references to socket locks, ultimately preventing us from
unblocking threads blocked on them.

To handle that situation, ensure that threads block on the socket lock
in steps, checking the network stack reset state at every step through
`LockGuard` "conditions". If a network stack reset is detected in this
way, threads will bail out.

Note that the "condition" lambda is necessary here, since socket locks
are allocated on a caller capability which we cannot heap-free-all. (see
recent additions to the `LockGuard` class).

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@ubc.ca>