Make `threadEntryGuard` not reset critical.
By moving the reset of `threadEntryGuard` at a different place in the
execution flow, we can remove it from the set of reset variables.

The idea here is to only reset `threadEntryGuard` in the case of a
crash, and only if the crash was triggered by a user thread, since
resetting is not needed in the case of a network thread crash (due to
deterministic execution flow, see comment in the code).

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@ubc.ca>